requirements.txt
# Core dependencies
streamlit
streamlit-nested-layout
streamlit-extras
# Data processing and visualization
pandas
altair
openpyxl
# Database
psycopg2-binary
sqlalchemy
# API and configuration
python-dotenv
google-generativeai